+++ changelog.md
@@ -2,6 +2,18 @@
 Visual Tracker (vistrack) change log
 James Jun, Flatiron Institute
 
+## [0.4.1] - 2019-08-19
+### Fixed
+- Bug fixed exporting `_posture.csv`
+  - `_posture.csv` table contains following five feature points (from head to tail direction)
+    '  Columns: x1(m), y1(m), x2(m), y2(m), x3(m), y3(m), x4(m), y4(m), x5(m), y5(m)',
+    '    x1(m): x coordinate of the head tip @ grid frame of reference',
+    '    y1(m): y coordinate of the head tip @ grid frame of reference',
+    '    x2(m): x coordinate of the head-mid section @ grid frame of reference',
+    '    x3(m): x coordinate of the mid section @ grid frame of reference',
+    '    x4(m): x coordinate of the mid-tail section @ grid frame of reference',
+    '    x5(m): x coordinate of the tail tip @ grid frame of reference',    
+
 ## [0.4.0] - 2019-08-19
 ### Added
 - `_posture.csv` and `_angles.csv` are exported when `Export CSV` button is pressed.